+++ NBE_5_7_2_02TEAM/src/main/java/io/twogether/nbe_5_7_2_02team/member/dto/response/LoginResponse.java
@@ -1,7 +1,7 @@
 package io.twogether.nbe_5_7_2_02team.member.dto.response;
 
 import io.twogether.nbe_5_7_2_02team.member.domain.Role;
-import io.twogether.nbe_5_7_2_02team.oauth.dto.TokenPair;
+import io.twogether.nbe_5_7_2_02team.oauth.dto.common.TokenPair;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
