+++ NBE_5_7_2_02TEAM/src/main/java/io/twogether/nbe_5_7_2_02team/oauth/dto/common/TokenPair.java
@@ -1,4 +1,4 @@
-package io.twogether.nbe_5_7_2_02team.oauth.dto;
+package io.twogether.nbe_5_7_2_02team.oauth.dto.common;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
