+++ NBE_5_7_2_02TEAM/src/main/java/io/twogether/nbe_5_7_2_02team/oauth/jwt/MemberDetailsFactory.java
@@ -2,7 +2,7 @@ package io.twogether.nbe_5_7_2_02team.oauth.jwt;
 
 import io.twogether.nbe_5_7_2_02team.global.exception.ErrorException;
 import io.twogether.nbe_5_7_2_02team.global.response.error.ErrorCode;
-import io.twogether.nbe_5_7_2_02team.oauth.dto.MemberDetails;
+import io.twogether.nbe_5_7_2_02team.oauth.dto.common.MemberDetails;
 
 import org.springframework.security.oauth2.core.user.OAuth2User;
