+++ NBE_5_7_2_02TEAM/src/main/java/io/twogether/nbe_5_7_2_02team/oauth/service/OAuthService.java
@@ -8,8 +8,8 @@ import io.twogether.nbe_5_7_2_02team.member.domain.Role;
 import io.twogether.nbe_5_7_2_02team.member.dto.request.SignUpRequest;
 import io.twogether.nbe_5_7_2_02team.member.dto.response.LoginResponse;
 import io.twogether.nbe_5_7_2_02team.member.dto.response.SignUpResponse;
-import io.twogether.nbe_5_7_2_02team.oauth.dto.MemberDetails;
-import io.twogether.nbe_5_7_2_02team.oauth.dto.TokenPair;
+import io.twogether.nbe_5_7_2_02team.oauth.dto.common.MemberDetails;
+import io.twogether.nbe_5_7_2_02team.oauth.dto.common.TokenPair;
 import io.twogether.nbe_5_7_2_02team.oauth.dto.response.GitHubLoginResponse;
 import io.twogether.nbe_5_7_2_02team.oauth.dto.response.GitHubUserInfoResponse;
 import io.twogether.nbe_5_7_2_02team.oauth.jwt.JwtTokenProvider;
